fix: removed EMPTY, added getAggregatedReigstryData()
